Revert PAR-2 scoring in costs for Task 2a (SAT)
When considering the runtime cost of instances with a timeout in the SAT
Competition, we should (and now we do) use the competition timeout itself,
not the PAR-2 score (including the penalty). This change only affects
queries with a timeout > 5000 s (which we did not conduct in our demo
notebook anyway).